+++ example/source/components/UI/Form.jsx
@@ -1,4 +1,3 @@
 import { createForm } from 'react-form-validation';
-import validator from '../../validator';
 
-export default createForm(validator);
+export default createForm();