+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+- Add onChange validation
+- Improve perfomance
+- Publish on GitHub Pages
 
 ## [0.2.1] - 2018-09-14
 - Fix main in package.json